backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
numpy==2.2.2
opencv-python==4.12.0.88
onnxruntime-gpu==1.22.0
pydantic==2.11.7
requests>=2.31.0
onnx>=1.15.0
psutil>=5.9.0
python-dotenv>=1.0.0


pytest>=8.0.0
pytest-cov>=6.0.0
hypothesis>=6.100.0
httpx>=0.28.0


scipy>=1.10.0
python-multipart>=0.0.6
python-docx>=1.1.0
openpyxl>=3.1.2
pillow>=10.2.0
ultralytics>=8.0.0


docxtpl>=0.17.0
pymupdf>=1.23.0
pywin32>=306; platform_system == 'Windows'
rio-tiler>=6.2.0
pyproj>=3.6.0
